Renaming wtprompt to lowercase everywhere
